Rename "sign up" -> "register"
Rename and move all routes to /register instead of /sign-up. More in line with what you're actually doing!
