+++ redwood/web/src/pages/Register/AllowCameraPage/AllowCameraPage.stories.tsx
@@ -1,0 +1,13 @@
+import {StoryMocks} from 'src/lib/StoryMocks'
+import RegisterLayout from '../RegisterLayout'
+import AllowCameraPage from './AllowCameraPage'
+
+export const Page = () => (
+  <StoryMocks user={{ethereumAddress: '0x123'}}>
+    <RegisterLayout>
+      <AllowCameraPage />
+    </RegisterLayout>
+  </StoryMocks>
+)
+
+export default {title: 'Pages/Register/3. Allow Camera'}
