+++ redwood/web/src/pages/Register/ConnectWalletPage/ConnectWalletPage.stories.tsx
@@ -1,0 +1,10 @@
+import RegisterLayout from '../RegisterLayout'
+import ConnectWalletPage from './ConnectWalletPage'
+
+export const Page = () => (
+  <RegisterLayout>
+    <ConnectWalletPage />
+  </RegisterLayout>
+)
+
+export default {title: 'Pages/Register/2. Connect Wallet'}
