+++ redwood/web/src/pages/Register/IntroPage/IntroPage.stories.tsx
@@ -1,0 +1,10 @@
+import RegisterLayout from '../RegisterLayout'
+import IntroPage from './IntroPage'
+
+export const Page = () => (
+  <RegisterLayout>
+    <IntroPage />
+  </RegisterLayout>
+)
+
+export default {title: 'Pages/Register/1. Intro'}
