+++ redwood/web/src/pages/Register/RegisterLogo.tsx
@@ -1,0 +1,10 @@
+import {Box} from '@chakra-ui/react'
+import LogoSVG from 'src/logo.svg'
+
+export default function RegisterLogo() {
+  return (
+    <Box mx="auto">
+      <LogoSVG height="100" width="100" />
+    </Box>
+  )
+}
