+++ redwood/web/src/pages/Register/SelfSubmitPage/SelfSubmitPage.stories.tsx
@@ -1,0 +1,10 @@
+import RegisterLayout from '../RegisterLayout'
+import SelfSubmitPage from './SelfSubmitPage'
+
+export const Page = () => (
+  <RegisterLayout>
+    <SelfSubmitPage />
+  </RegisterLayout>
+)
+
+export default {title: 'Pages/Register/Self Submit'}
