--- redwood/web/src/pages/SignUp/AllowCameraPage/AllowCameraPage.stories.tsx
@@ -1,13 +1,0 @@
-import {StoryMocks} from 'src/lib/StoryMocks'
-import SignUpLayout from '../SignUpLayout'
-import AllowCameraPage from './AllowCameraPage'
-
-export const Page = () => (
-  <StoryMocks>
-    <SignUpLayout>
-      <AllowCameraPage />
-    </SignUpLayout>
-  </StoryMocks>
-)
-
-export default {title: 'Pages/Sign Up/3. Allow Camera'}
