--- redwood/web/src/pages/SignUp/ConnectWalletPage/ConnectWalletPage.stories.tsx
@@ -1,10 +1,0 @@
-import SignUpLayout from '../SignUpLayout'
-import ConnectWalletPage from './ConnectWalletPage'
-
-export const Page = () => (
-  <SignUpLayout>
-    <ConnectWalletPage />
-  </SignUpLayout>
-)
-
-export default {title: 'Pages/Sign Up/2. Connect Wallet'}
