--- redwood/web/src/pages/SignUp/IntroPage/IntroPage.stories.tsx
@@ -1,10 +1,0 @@
-import SignUpLayout from '../SignUpLayout'
-import IntroPage from './IntroPage'
-
-export const Page = () => (
-  <SignUpLayout>
-    <IntroPage />
-  </SignUpLayout>
-)
-
-export default {title: 'Pages/Sign Up/1. Intro'}
