--- redwood/web/src/pages/SignUp/SelfSubmitPage/SelfSubmitPage.stories.tsx
@@ -1,10 +1,0 @@
-import SignUpLayout from '../SignUpLayout'
-import SelfSubmitPage from './SelfSubmitPage'
-
-export const Page = () => (
-  <SignUpLayout>
-    <SelfSubmitPage />
-  </SignUpLayout>
-)
-
-export default {title: 'Pages/Sign Up/Self Submit'}
