--- redwood/web/src/pages/SignUp/SignUpLogo.tsx
@@ -1,10 +1,0 @@
-import {Box} from '@chakra-ui/react'
-import LogoSVG from 'src/logo.svg'
-
-export default function SignUpLogo() {
-  return (
-    <Box mx="auto">
-      <LogoSVG height="100" width="100" />
-    </Box>
-  )
-}
